pm.cc: Implement run_man and write_to_tempfile
Now man(1) invocation is integrated into the program, and we are also
able to write the HTML output to a temp file and print its path so that
we can independently verify that the HTML output does work.

The next step is to get rid of write_to_tempfile, implement auto
refresh, and write to a socket instead. Then we'll be able to
autorefresh by listening on a local port.