pm.cc: Fix child process aborting bug
Previously, on execvp error, we throw an uncaught PMException within the
server-spawning child process which runs in the server controller
thread, which is apparently a bad idea. In fact, even exitting there
won't help, because exitting anything other than the main thread leads
to child processes spawned in the thread being terminated with SIGABRT
instead of exitting with the appropriate exit status. Therefore, we
employ the trick of executing `bin/sh -c "exit 127"' which is guaranteed
to make the child process exit with status 127 even within a thread.